feat(pie): empirical score pool + soft counting for ranking stability
Multi-role analysis identified two sources of variance killing accuracy:

1. SOFT COUNTING in track records (types/pie.ts, engine.ts):
   exact[i] was incremented via a random pool sample (wSamplePool) —
   making player rankings non-deterministic and noisy. Now exact[i] gets
   the actual pool probability of the real score (e.w / sPool.total).
   Swaps Int32Array → Float32Array for the exact field. Result: stable,
   deterministic rankings so the real best players always surface.

2. EMPIRICAL SCORE POOL (engine.ts):
   Score prediction now uses the actual distribution of scorelines from
   the current tournament's training matches (LOO-safe). Real prode
   players in a 100K competition learn from watching the same matches
   and cluster their bets on scores they've seen. The model now does the
   same: buildEmpiricalScorePool counts home/draw/away results from
   wcResults (the training set) and picks the modal score.
   Requires ≥5 examples per outcome type; falls back to leader sample
   when data is sparse (early tournament).

PIECard updated: exactCorrect display rounds to 1 decimal since it's
now a float; threshold for highlighting changed to ≥0.5.

+++ migration/src/types/pie.ts
@@ -36,8 +36,8 @@ export interface PIEResult {
 
 // Track records for 1M players (typed arrays, exported for hook memoization)
 export interface PIETrackRecords {
-  correct:  Int32Array;   // correct winner picks per player
-  exact:    Int32Array;   // correct exact scores per player
-  total:    Int32Array;   // total matches played
-  upset:    Int32Array;   // upset picks correct per player
+  correct:  Int32Array;    // correct winner picks per player
+  exact:    Float32Array;  // expected exact score credit per player (soft counting — pool probability, not a sample)
+  total:    Int32Array;    // total matches played
+  upset:    Int32Array;    // upset picks correct per player
 }
